Join refactoring: addOsnToChannel function put into FabricNetworkManager.
